+++ core/node.h
@@ -21,8 +21,8 @@ namespace oidn {
 
     virtual Ref<Tensor> getDst() const { return nullptr; }
 
-    virtual size_t getScratchpadSize() const { return 0; }
-    virtual void setScratchpad(const Ref<Tensor>& scratchpad) {}
+    virtual size_t getScratchSize() const { return 0; }
+    virtual void setScratch(const Ref<Tensor>& scratch) {}
 
     __forceinline Device* getDevice() { return device.get(); }
   };
@@ -35,12 +35,12 @@ namespace oidn {
   protected:
     dnnl::primitive prim;
     std::unordered_map<int, dnnl::memory> args;
-    Ref<Tensor> scratchpad;
+    Ref<Tensor> scratch;
 
   public:
     DNNLNode(const Ref<Device>& device) : Node(device) {}
 
-    size_t getScratchpadSize() const override
+    size_t getScratchSize() const override
     {
       const auto primDesc = prim.get_primitive_desc();
       const dnnl_memory_desc_t* scratchpadDesc = dnnl_primitive_desc_query_md(primDesc, dnnl_query_scratchpad_md, 0);
@@ -49,10 +49,10 @@ namespace oidn {
       return dnnl_memory_desc_get_size(scratchpadDesc);
     }
 
-    void setScratchpad(const Ref<Tensor>& scratchpad) override
+    void setScratch(const Ref<Tensor>& scratch) override
     {
-      this->scratchpad = scratchpad;
-      args.insert(std::make_pair(DNNL_ARG_SCRATCHPAD, scratchpad->mem));
+      this->scratch = scratch;
+      args.insert(std::make_pair(DNNL_ARG_SCRATCHPAD, scratch->mem));
     }
 
     void execute() override
@@ -67,9 +67,7 @@ namespace oidn {
   class BNNSNode : public Node
   {
   protected:
-    BNNSFilter  filter = nullptr;
-    const void* inPtr  = nullptr;
-    void*       outPtr = nullptr;
+    BNNSFilter filter = nullptr;
 
   public:
     BNNSNode(const Ref<Device>& device) : Node(device) {}
@@ -78,11 +76,6 @@ namespace oidn {
     {
       if (filter)
         BNNSFilterDestroy(filter);
-    }
-
-    void execute() override
-    {
-      BNNSFilterApply(filter, inPtr, outPtr);
     }
   };